backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
google-generativeai
pillow
pydantic